pyproject.toml
[tool.poetry]
name = "storage-api"
version = "0.1.0"
description = ""
authors = ["Jorique <guselnikovgeorgiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.115.0"}
uvicorn = "^0.31.0"
sqlalchemy = "^2.0.35"
asyncpg = "^0.29.0"
alembic = "^1.13.3"
pydantic-settings = "^2.5.2"
ipython = "^8.27.0"
httpx = "^0.27.2"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
gunicorn = "^23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
target-version = ["py312"]

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
remove_all_unused_imports = true